refactor: pull out all NTLM handling out of `Login7Payload`
`Login7Payload` should not know anything about NTLM, it only cares about 
getting a SSPI blob.
