+++ test/unit/login7-payload-test.js
@@ -74,84 +74,6 @@ exports.create = function(test) {
   test.done();
 };
 
-exports.createNTLM = function(test) {
-  var loginData = {
-    userName: 'user',
-    password: 'pw',
-    appName: 'app',
-    serverName: 'server',
-    domain: 'domain',
-    workstation: 'workstation',
-    language: 'lang',
-    database: 'db',
-    packetSize: 1024,
-    tdsVersion: '7_2'
-  };
-
-  var payload = new Login7Payload(loginData);
-
-  var expectedLength =
-    4 + // Length
-    32 + // Variable
-    2 +
-    2 +
-    2 * payload.hostname.length +
-    2 +
-    2 +
-    2 * 0 +
-    2 +
-    2 +
-    2 * 0 +
-    2 +
-    2 +
-    2 * loginData.appName.length +
-    2 +
-    2 +
-    2 * loginData.serverName.length +
-    2 +
-    2 +
-    2 * 0 + // Reserved
-    2 +
-    2 +
-    2 * payload.libraryName.length +
-    2 +
-    2 +
-    2 * loginData.language.length +
-    2 +
-    2 +
-    2 * loginData.database.length +
-    payload.clientId.length +
-    2 +
-    2 +
-    payload.sspi.length + // NTLM
-    2 +
-    2 +
-    2 * payload.attachDbFile.length +
-    2 +
-    2 +
-    2 * payload.changePassword.length +
-    4; // cbSSPILong
-
-  const data = payload.toBuffer();
-  test.strictEqual(data.length, expectedLength);
-
-  var protocolHeader = payload.sspi.slice(0, 8).toString('utf8');
-  test.strictEqual(protocolHeader, 'NTLMSSP\u0000');
-
-  var workstationName = payload.sspi
-    .slice(payload.sspi.length - 17)
-    .toString('ascii')
-    .substr(0, 11);
-  test.strictEqual(workstationName, 'WORKSTATION');
-
-  var domainName = payload.sspi
-    .slice(payload.sspi.length - 6)
-    .toString('ascii');
-  test.strictEqual(domainName, 'DOMAIN');
-
-  test.done();
-};
-
 exports.createSSPI = function(test) {
   var loginData = {
     userName: '',
